Move playlist json to its own file.
